+++ diesel_attribute_parser/src/parsers/postgres_type.rs
@@ -3,7 +3,8 @@ use syn::punctuated::Punctuated;
 use syn::token::Comma;
 use syn::{Ident, LitInt, LitStr};
 
-use crate::util::{POSTGRES_TYPE_NOTE, POSTGRES_TYPE_NOTE_ID, parse_eq, unknown_attribute};
+use crate::notes::{POSTGRES_TYPE_NOTE, POSTGRES_TYPE_NOTE_ID};
+use crate::util::{parse_eq, unknown_attribute};
 
 enum Attr {
     Oid(Ident, LitInt),
@@ -34,6 +35,7 @@ impl Parse for Attr {
     }
 }
 
+#[derive(Debug)]
 pub enum PostgresType {
     Fixed(LitInt, LitInt),
     Lookup(LitStr, Option<LitStr>),
