+++ diesel_derives/Cargo.toml
@@ -16,6 +16,7 @@ syn = { version = "2.0", features = ["derive", "fold", "full", "visit-mut"] }
 quote = "1.0.9"
 proc-macro2 = "1.0.27"
 diesel_table_macro_syntax = {version = "0.3", path = "../diesel_table_macro_syntax"}
+diesel_attribute_parser = { path = "../diesel_attribute_parser" }
 dsl_auto_type = { version = "0.2", path = "../dsl_auto_type" }
 
 [dev-dependencies]
