+++ diesel_derives/src/sql_type.rs
@@ -4,8 +4,8 @@ use syn::Result;
 use syn::{DeriveInput, Ident};
 
 use crate::model::Model;
-use crate::parsers::PostgresType;
 use crate::util::wrap_in_dummy_mod;
+use diesel_attribute_parser::parsers::PostgresType;
 
 pub fn derive(item: DeriveInput) -> Result<TokenStream> {
     let model = Model::from_item(&item, true, false)?;
